refactor: betterness for fuseinfo pass goodness
