+++ jaclang/compiler/absyntree.py
@@ -6,20 +6,29 @@ import ast as ast3
 import os
 from hashlib import md5
 from types import EllipsisType
-from typing import Any, Callable, Generic, Optional, Sequence, Type, TypeVar
+from typing import (
+    Any,
+    Callable,
+    Generic,
+    Optional,
+    Sequence,
+    TYPE_CHECKING,
+    Type,
+    TypeVar,
+)
 
 from jaclang.compiler import TOKEN_MAP
 from jaclang.compiler.codeloc import CodeGenTarget, CodeLocInfo
-from jaclang.compiler.constant import Constants as Con, EdgeDir
-from jaclang.compiler.constant import DELIM_MAP, Tokens as Tok
+from jaclang.compiler.constant import Constants as Con, EdgeDir, SymbolType
+from jaclang.compiler.constant import DELIM_MAP, SymbolAccess, Tokens as Tok
 from jaclang.compiler.semtable import SemRegistry
-from jaclang.compiler.symtable import (
-    Symbol,
-    SymbolAccess,
-    SymbolTable,
-    SymbolType,
-)
 from jaclang.utils.treeprinter import dotgen_ast_tree, print_ast_tree
+
+if TYPE_CHECKING:
+    from jaclang.compiler.symtable import (
+        Symbol,
+        SymbolTable,
+    )
 
 
 class AstNode:
