+++ jaclang/compiler/passes/main/access_modifier_pass.py
@@ -8,6 +8,7 @@ from typing import Optional
 
 import jaclang.compiler.absyntree as ast
 from jaclang.compiler.passes.main.sym_tab_build_pass import SymTabPass, SymbolAccess
+from jaclang.compiler.symtable import SymbolTable
 from jaclang.settings import settings
 
 
@@ -28,7 +29,7 @@ class AccessCheckPass(SymTabPass):
         """Access check."""
         node_info = (
             node.sym_tab.lookup(node.sym_name)
-            if isinstance(node.sym_tab, ast.SymbolTable)
+            if isinstance(node.sym_tab, SymbolTable)
             else None
         )
